+++ src/store/actions.js
@@ -9,6 +9,8 @@ const addNotification = function({ commit, state }, val) {
 
   if (!Array.isArray(newNotif[val[0]])) newNotif[val[0]] = [];
 
+  if (!Array.isArray(newNotif[val[0]])) newNotif[val[0]] = []
+
   newNotif[val[0]].push({
     title: val[2],
     read: false,
@@ -91,6 +93,7 @@ const updatePageState = function({ commit }, arr) {
 
 export default {
   addNotification,
+  addCustomPath,
   checkIfOnline,
   clearWallet,
   createAndSignTx,