+++ src/wallets/hardware/secalot/index.js
@@ -1,0 +1,5 @@
+import SecalotWallet from './secalotWallet'
+
+export {
+  SecalotWallet
+}