+++ src/wallets/software/mnemonic/index.js
@@ -1,0 +1,5 @@
+import MnemonicWallet from './mnemonicWallet'
+
+export {
+  MnemonicWallet
+}